WI-001708: clear the custom flag on the whole Pricing Proposal chain
Found while working out what could be tested: all four chain doctypes carry
custom = 1 in the database, because they were built through the UI in production.
Frappe returns the base Document class for a custom doctype, so their controllers
never load - WI-001707's "only one enabled" and unique-effective-from checks, and
WI-001713's Budget Configuration resolution, all silently did nothing. The
features would have tested as broken while the code was fine.

The take-ownership patch added for the two missing masters is broadened to cover
Budget Configuration, Pricing Proposal, Manpower Service Item and Pricing Proposal
Service Item, and renamed accordingly. Same pre_model_sync approach as WI-001576.
It only writes when the flag is actually set, and is a no-op for any doctype not
present yet.

Confirmed against the database: Service Category and Service Role resolve to their
controller classes after the earlier run of this patch, while the other four
resolved to plain Document.
